% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_join.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Join and Coalesce}
\usage{
coalesce_join(x, y, by, join = full_join)
}
\arguments{
\item{x}{first data frame}

\item{y}{second data frame}

\item{by}{columns to join by}

\item{join}{type of join}
}
\description{
Join two dfs and then combine duplicate columns by replacing NAs in x
with y
}
