% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_threats.R
\name{recode_threat}
\alias{recode_threat}
\title{Convert threats from text to codes used in database}
\usage{
recode_threat(x, var, threat_code = threat_code)
}
\arguments{
\item{x}{A character vector with the threat level}

\item{var}{which threat component is being converted}

\item{threat_code}{A data.frame with corresponding threat levels and codes}
}
\value{
a data.frame
}
\description{
If a range of threat levels is given then the average of the two threat codes
is used. E.g. a threat level of "Extreme - Moderate (11-100\%)" would get a
code of 4 which is the average of 3 (Moderate) and 5 (Extreme).
}
